pyproject.toml
[project]
name = "nexo-server"
version = "0.1.0"
description = "Nexo backend service"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "argon2-cffi>=23.1.0",
  "asyncpg>=0.29.0",
  "boto3>=1.36.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.6",
  "httpx>=0.28.1",
  "prometheus-client>=0.21.1",
  "pydantic-settings>=2.7.1",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.36",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.34.0",
  "bleach>=6.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0",
  "testcontainers[postgres,redis]>=4.9.1",
  "ruff>=0.8.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[tool.ruff]
line-length = 100
target-version = "py312"
